requirements.txt
python-dotenv
pandas
numpy
scikit-learn
pymongo
pymongo[srv]==3.11
certifi
pyyaml
mlflow
# -e .